requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
Django==2.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-redis==4.12.1
django-simpleui==2021.3
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
dnspython==1.16.0
eventlet==0.30.2
greenlet==1.0.0
idna==2.10
importlib-metadata==3.7.2
itypes==1.2.0
Jinja2==2.11.3
kombu==5.0.2
MarkupSafe==1.1.1
Pillow==8.1.2
prompt-toolkit==3.0.16
pycparser==2.20
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
python-alipay-sdk==3.0.1
pytz==2021.1
qcloudsms-py==0.1.4
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
